+++ app/controllers/api/v1_controller.rb
@@ -4,6 +4,8 @@ class Api::V1Controller < ApplicationController
   rescue_from ActiveRecord::RecordNotFound, with: :render_404
   skip_before_action :verify_authenticity_token
 
+  include UrlHelpers
+
   private
 
   #
