+++ app/controllers/concerns/annotations.rb
@@ -1,6 +1,5 @@
 module Annotations 
 	extend ActiveSupport::Concern
-  include Rails.application.routes.url_helpers
 
 	def common_annotations(extension, version, asset=nil)
     tags = if asset.present? && asset.labels.present? 
@@ -9,6 +8,7 @@ module Annotations
       version.release_assets.map{|a| a.labels}.flatten.reject(&:blank?).uniq.sort.join(', ')
     end
     tags ||= nil
+
     annotations = {
       'io.sensio.bonsai.url' => owner_scoped_extension_url(extension),
       'io.sensio.bonsai.api_url' => owner_scoped_release_asset_builds_api_v1_url(extension, version),
